requirements.txt
Flask==2.2.2
instagram-basic-display==1.0.5
celery==5.2.7
flower==1.2.0
redis==4.3.4
requests==2.28.1
waitress==2.1.2
tweepy==4.10.1
sqlalchemy==1.4.41
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.1
python-dotenv==0.21.0

sentry-sdk[flask]==1.9.10
demjson==2.2.4
pycryptodome==3.15.0